+++ changelog.md
@@ -1,11 +1,16 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.1.4] - May 23, 2022
+- Removed the template folder and started fetching the templates from github.
+- Moved all flavours to one new repo called freshenv-flavours.
+- Exception handling in the build command.
+
 ## [1.1.2] - May 23, 2022
-- fixed the manifest file to add the template/simple dockerfile
+- fixed the manifest file to add the template/simple dockerfile.
 
 ## [1.1.1] - May 23, 2022
-- fixed the manifest file to add the template/simple dockerfile
+- fixed the manifest file to add the template/simple dockerfile.
 
 ## [1.1.0] - May 21, 2022
 - Enabled custom flavours by adding build command.